requirements.txt
azure-storage-blob
azure-search-documents
azure-ai-formrecognizer
flask
python-dotenv
requests
werkzeug
pytz
